be_hydroponic_system/pyproject.toml
[tool.poetry]
name = "luna"
version = "0.1.0"
description = ""
authors = ["kajtek.k.janusz@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
djangorestframework = "^3.15.2"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
